productivity_software/requirements.txt
django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
pyotp==2.9.0
Pillow==10.1.0
drf-spectacular==0.26.5
pytest==7.4.0
pytest-django==4.5.2

# Database dependencies - essential
psycopg2-binary==2.9.9  # PostgreSQL adapter
redis==5.0.1  # Redis Python client
django-redis==5.4.0  # Redis cache backend for Django

# Database dependencies - pending resolution of dependency conflicts
# MongoDB dependencies
# pymongo==4.6.1  # MongoDB Python driver
# git+https://github.com/doableware/djongo.git  # MongoDB connector for Django

# Elasticsearch dependencies
# elasticsearch==8.10.0  # Elasticsearch Python client
# django-elasticsearch-dsl==8.0  # Elasticsearch integration for Django 

# Production dependencies
gunicorn==21.2.0
gevent==23.9.1
python-dotenv==1.0.0
django-storages==1.14.2
sentry-sdk==1.39.1
# Security enhancements
django-csp==3.7
django-permissions-policy==4.17.0
whitenoise==6.6.0 